Move the tests and mode to account for name collisions and use PyTorch testing infra.
