+++ test/complextensor/test_complextensor.py
@@ -1,24 +1,40 @@
 # Owner(s): ["module: complex"]
 from __future__ import annotations
 
-from typing import Any, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
 import torch
-import torch.distributed as dist
-from torch.complex.complextensor.ops.common import (
-    _as_complex_tensor,
-    _get_op_name,
-    COMPLEX_OPS_TABLE,
-    ComplexTensorMode,
-    FORCE_TEST_LIST,
-    OpOverloadPacket,
-)
+from torch.testing._internal.common_distributed import MultiProcessTestCase
+
+
+# Support both when imported from elsewhere or directly as a file
+try:
+    from .utils import (
+        COMPLEX_DTYPES,
+        Descriptor,
+        force_test_op_db,
+        get_overload_packet_from_name,
+        implemented_op_db,
+        TestCase,
+        Variant,
+    )
+except ImportError:
+    from utils import (
+        COMPLEX_DTYPES,
+        Descriptor,
+        force_test_op_db,
+        get_overload_packet_from_name,
+        implemented_op_db,
+        TestCase,
+        Variant,
+    )
+
+from torch.complextensor.ops.common import ComplexTensorMode
 from torch.testing._internal.common_device_type import (
     instantiate_device_type_tests,
     OpDTypes,
     ops,
 )
-from torch.testing._internal.common_methods_invocations import op_db
 from torch.testing._internal.common_utils import (
     parametrize,
     run_tests,
@@ -26,61 +42,11 @@ from torch.testing._internal.common_utils import (
     unMarkDynamoStrictTest,
 )
 
-from .utils import COMPLEX_DTYPES, Descriptor, TestCase, Variant
-
 
 if TYPE_CHECKING:
     from torch.testing._internal.opinfo.core import OpInfo
 
-
-torch._dynamo.config.recompile_limit = float("inf")
-torch._dynamo.config.accumulated_recompile_limit = float("inf")
-
 aten = torch.ops.aten
-
-complex_op_db = tuple(
-    filter(lambda op: any(op.supports_dtype(ct, "cpu") for ct in COMPLEX_DTYPES), op_db)
-)
-
-
-def get_overload_packet_from_name(name: str) -> OpOverloadPacket:
-    for domain_name in torch.ops:
-        op_namespace = getattr(torch.ops, domain_name)
-        op: OpOverloadPacket | None = getattr(op_namespace, name, None)
-        if op is not None:
-            return op
-
-    raise RuntimeError(f"No op with {name=} found.")
-
-
-force_test_names = set(map(_get_op_name, FORCE_TEST_LIST))
-implemented_op_names = (
-    set(map(_get_op_name, COMPLEX_OPS_TABLE.keys())) - force_test_names
-)
-implemented_op_db = tuple(
-    filter(lambda op: op.name in implemented_op_names, complex_op_db)
-)
-force_test_op_db = tuple(filter(lambda op: op.name in force_test_names, op_db))
-
-tested_op_names = {op.name for op in implemented_op_db} | {
-    op.name for op in force_test_op_db
-}
-non_tested_ops = {
-    op for op in COMPLEX_OPS_TABLE if _get_op_name(op) not in tested_op_names
-}
-
-if len(non_tested_ops) != 0:
-    import textwrap
-    import warnings
-
-    list_missing_ops = "\n".join(sorted([str(op) for op in non_tested_ops]))
-    warnings.warn(
-        "Not all implemented ops are tested. List of ops missing tests:"
-        f"\n{textwrap.indent(list_missing_ops, '    ')}",
-        UserWarning,
-        stacklevel=2,
-    )
-
 
 SKIPS = {
     Descriptor(op=aten.empty_like, variant=None): "Non-deterministic output",
@@ -215,25 +181,6 @@ EXTRA_KWARGS = {
     },
 }
 
-STORE = dist.HashStore()
-dist.init_process_group(store=STORE, rank=0, world_size=1)
-DEVICE_MESH = dist.init_device_mesh("cpu", mesh_shape=(1,))
-
-
-def _as_complex_dtensor(arg: torch.Tensor | Any) -> torch.Tensor | Any:
-    if not isinstance(arg, torch.Tensor):
-        return arg
-
-    return dist.tensor.DTensor.from_local(
-        _as_complex_tensor(arg), device_mesh=DEVICE_MESH
-    )
-
-
-TRANSFORM_FUNCS = {
-    Variant.Op: _as_complex_tensor,
-    Variant.Distributed: _as_complex_dtensor,
-}
-
 
 class TestComplexTensor(TestCase):
     _default_dtype_check_enabled = True
@@ -252,58 +199,18 @@ class TestComplexTensor(TestCase):
     def test_maybe_error(self, device, dtype, op: OpInfo, compile: bool):
         self.check_consistency(device, dtype, op, compile, Variant.Op)
 
+
+@unMarkDynamoStrictTest
+class TestComplexDistributed(TestCase, MultiProcessTestCase):
     @ops(implemented_op_db, allowed_dtypes=list(COMPLEX_DTYPES))
     def test_distributed(self, device, dtype, op: OpInfo):
         self.check_consistency(device, dtype, op, False, Variant.Distributed)
 
-    def check_consistency(
-        self, device: torch.device, dtype, op: OpInfo, compile: bool, variant: Variant
-    ) -> None:
-        test_info = Descriptor(
-            op=get_overload_packet_from_name(op.name),
-            device=device,
-            dtype=dtype,
-            compile=compile,
-            variant=variant,
-        )
-        for xfail_info, reason in SKIPS.items():
-            if xfail_info.matches(test_info):
-                self.skipTest(reason)
-
-        kwargs = {}
-        for extra_info, extra_kw in EXTRA_KWARGS.items():
-            if extra_info.matches(test_info):
-                kwargs = extra_kw
-                break
-
-        sample_inputs = op.sample_inputs(device, dtype)
-        op_eager = op
-        if compile:
-            op = torch.compile(op, fullgraph=True)
-
-        transform_fn = TRANSFORM_FUNCS[variant]
-
-        for sample_input in sample_inputs:
-
-            def expected(sample_input=sample_input):
-                return op_eager(
-                    sample_input.input, *sample_input.args, **sample_input.kwargs
-                )
-
-            subclass_sample = sample_input.transform(transform_fn)
-
-            def actual(subclass_sample=subclass_sample):
-                return op(
-                    subclass_sample.input,
-                    *subclass_sample.args,
-                    **subclass_sample.kwargs,
-                )
-
-            self.assertSameResult(expected, actual, ignore_exc_types=compile, **kwargs)
-
 
 @unMarkDynamoStrictTest
 class TestComplexBwdGradients(TestGradients):
+    _default_dtype_check_enabled = True
+
     @ops(
         implemented_op_db,
         dtypes=OpDTypes.supported_backward,
@@ -333,6 +240,7 @@ class TestComplexBwdGradients(TestGradients):
 
 instantiate_device_type_tests(TestComplexTensor, globals())
 instantiate_device_type_tests(TestComplexBwdGradients, globals())
+instantiate_device_type_tests(TestComplexDistributed, globals())
 
 if __name__ == "__main__":
     run_tests()
