--- torch/complex/__init__.py
@@ -1,4 +1,0 @@
-from .complextensor import ComplexTensor
-
-
-__all__ = ["ComplexTensor"]
